refactor(cache): extract cache into its own package
Move the memory/redis cache implementation out of internal/sqsnotify2
into a dedicated internal/cache package and export NewMemoryCache.
