+++ internal/cache/cache_redis.go
@@ -1,4 +1,4 @@
-package sqsnotify2
+package cache
 
 import (
 	"context"
